Move coffee details files to the pages folder
