+++ src/App.js
@@ -2,8 +2,7 @@ import React, { Component } from 'react'
 import { Redirect, Route, Switch } from 'react-router-dom'
 import { Container } from 'semantic-ui-react'
 import { graphql } from 'react-apollo'
-import LeanCoffeeDetails from './components/coffees/Details/Details'
-import { ListLeanCoffees } from './pages'
+import { LeanCoffeeDetails, ListLeanCoffees } from './pages'
 import ListParticipants from './components/participants/List'
 import ListTopics from './components/topics/List'
 import Home from './components/Home'
