+++ src/pages/index.js
@@ -1,1 +1,2 @@
 export { default as ListLeanCoffees } from './coffee/list/List'
+export { default as LeanCoffeeDetails } from './coffee/details/Details'
